backend/requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
google-cloud-firestore
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic-settings
email-validator
pyasn1==0.4.8
pyasn1-modules==0.2.8
